+++ apps/demo/emails/02-Matte/activation.tsx
@@ -40,7 +40,7 @@ export const ActivationEmail = ({ companyName, url }: ActivationEmailProps) => (
         <Container className="mx-auto max-w-[640px] px-4 pt-16 pb-6">
           <Section className="shadow-collage-card rounded-[8px]">
             <Section className="bg-bg border-stroke rounded-[8px] border">
-              <Section className="px-10 pt-16">
+              <Section className="mobile:px-6! px-10 pt-16">
                 <Img
                   src={`${baseUrl}/static/collage/collage-image-1.png`}
                   alt=""
@@ -50,7 +50,7 @@ export const ActivationEmail = ({ companyName, url }: ActivationEmailProps) => (
                 />
               </Section>
 
-              <Section className="px-10 pt-8">
+              <Section className="mobile:px-6! px-10 pt-8">
                 <Section className="mb-9">
                   <Text className="font-48 text-fg m-0 font-sans">
                     Almost there
@@ -71,7 +71,7 @@ export const ActivationEmail = ({ companyName, url }: ActivationEmailProps) => (
                 </Button>
               </Section>
 
-              <Section className="px-10 pt-16 pb-8">
+              <Section className="mobile:px-6! px-10 pt-16 pb-8">
                 <Text className="font-11 font-inter text-fg-3 m-0 max-w-[310px]">
                   If you didn&apos;t create an account, you can safely ignore
                   this email.