+++ packages/react-email/CHANGELOG.md
@@ -1,5 +1,7 @@
 # react-email
 
+## 6.0.0-canary.2
+
 ## 6.0.0-canary.1
 
 ### Patch Changes